+++ src/main/java/com/yahoo/sketches/characterization/uniquecount/ThetaUpdateSpeedProfile.java
@@ -52,7 +52,7 @@ public class ThetaUpdateSpeedProfile extends BaseUpdateSpeedProfile {
       sketch.update(++vIn);
     }
     final long updateTime_nS = System.nanoTime() - startUpdateTime_nS;
-    return updateTime_nS / uPerTrial;
+    return (double) updateTime_nS / uPerTrial;
   }
 
 }
